Save progress on HW2 CNN notebook
The training loop now runs. However, the accuracy doesn't improve. I
first want to make sure I'm calculating that correctly. Then I want to
play around with the model architecture and hyperparameters. It takes a
while to run through even a few epochs, so I may reduce the size of the
training and validation datasets until I find an approach that's more
promising.