Save progress on HW3 notebook
The model now improves while training on multicharacter input images
rather than the loss remaining essentially constant. The model still
isn't perfect. It seems to have trouble with the letter e in particular
and doesn't do a great job of splitting double letters. But it feels
like I'm now on the right path and only have to do a little tweaking to
get it where I want.